backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dateutil==2.8.2
werkzeug==3.0.1
